int changed for size_t

+++ src/guido/xml2guidovisitor.cpp
@@ -59,7 +59,7 @@ void xml2guidovisitor::flushHeader ( scoreHeader& header )
 	if (header.fTitle) {
 		Sguidoelement tag = guidotag::create("title");
 		string title = header.fTitle->getValue();
-		int pos = title.find ('"');
+		size_t pos = title.find ('"');
 		while (pos != string::npos) {
 			title = title.replace (pos, 1, "'");
 			pos = title.find ('"', pos);
